Possible perf improvement
I am not sure as to the performance improvements but for reading a long string, there will be a multiple allocations. This is because strings in Javascript are immutable and adding a character to a string causes a new allocation. Instead this change using an Array to temporary aggregate the characters and then at the end, join them together. There seems to be some change in benchmarks but they varied a lot.